add email validation, add owner service, refactor owner controller, refactor owner and globalentityproperties models, add findOwnerByEmail method
