+++ src/main/java/com/example/sports_participant_be_main/controllers/OwnerController.java
@@ -1,21 +1,21 @@
 package com.example.sports_participant_be_main.controllers;
 
 import com.example.sports_participant_be_main.models.Owner;
-import com.example.sports_participant_be_main.repositories.OwnerRepo;
+import com.example.sports_participant_be_main.services.OwnerService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
-import java.time.LocalDateTime;
+import javax.validation.Valid;
 
 @RestController
 @RequestMapping("/owner")
 @AllArgsConstructor
 public class OwnerController {
 
-    private final OwnerRepo ownerRepo;
+    private final OwnerService ownerService;
 
     @PostMapping("/add")
-    public void add(@RequestBody Owner owner){
-        ownerRepo.save(owner);
+    public void add(@Valid @RequestBody Owner owner){
+        ownerService.save(owner);
     }
 }
